New Controller
I did a huge update to the control system, it now works a lot better together and works fully for a curtailment of one source alone. However the combination of sources still creates problems in the curtailment, however no errors.
Also I cleaned up the code a little, yet still some remained, this is mostly due to the implementation of the new controller.